pyproject.toml
[project]
name = "sentinel-curator"
version = "0.1.0"
description = "Secure LLM-curated military asset tracking system with RBAC and PostGIS"
requires-python = ">=3.11"
authors = [
    { name = "Iain Reid", email = "iain.reid@example.com" }
]
license = { text = "RESTRICTED - Internal use only" }
readme = "README.md"

dependencies = [
    # Web framework
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    # Database
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",         # async PostgreSQL driver
    "psycopg2-binary>=2.9.9",  # sync driver for migrations/scripts
    "geoalchemy2>=0.14.0",     # PostGIS geometry types for SQLAlchemy
    # Data validation
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    # LLM / RAG
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-openai>=0.1.0",  # swappable — config-driven
    # Cryptography (FIPS 140-3 aligned)
    "cryptography>=42.0.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",           # async test client for FastAPI
    "black>=24.3.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
    "bandit>=1.7.8",
    "pip-audit>=2.7.0",
    "sqlalchemy[mypy]>=2.0.0",
]

[project.scripts]
curator = "sentinel_curator.api.main:run"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[tool.setuptools.packages.find]
where = ["src"]

# ---------------------------------------------------------------------------
# Black
# ---------------------------------------------------------------------------
[tool.black]
line-length = 88
target-version = ["py311"]

# ---------------------------------------------------------------------------
# Ruff (replaces flake8 + isort)
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "S", "UP"]
ignore = ["S101"]   # allow assert in tests

# ---------------------------------------------------------------------------
# Mypy
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy", "sqlalchemy.ext.mypy.plugin"]

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=sentinel_curator --cov-report=term-missing --cov-fail-under=70"

# ---------------------------------------------------------------------------
# Coverage
# ---------------------------------------------------------------------------
[tool.coverage.run]
source = ["src/sentinel_curator"]
omit = ["*/tests/*", "*/__init__.py"]

# ---------------------------------------------------------------------------
# Bandit (security linting)
# ---------------------------------------------------------------------------
[tool.bandit]
skips = ["B101"]   # assert_used — acceptable in test code
